requirements.txt
altgraph==0.17.3
black==22.10.0
click==8.1.3
colorama==0.4.6
future==0.18.2
mypy-extensions==0.4.3
pathspec==0.10.2
pdf2image==1.16.0
pefile==2022.5.30
Pillow==9.3.0
pip==21.3.1
platformdirs==2.5.4
pyinstaller-hooks-contrib==2022.13
pyinstaller==5.6.2
pywin32-ctypes==0.2.0
setuptools==60.2.0
tomli==2.0.1
wheel==0.37.1
